fixed merge issue bug

+++ datasets/data_loader.py
@@ -13,7 +13,7 @@ def load_preloaded(dataset_name, seed, training=True ,X_y=False):
 
     filename = f"TRAINING_{seed}_{dataset_name.upper()}.txt" if training else f"TEST_{seed}_{dataset_name.upper()}.txt"
 
-    df = pd.read_csv(os.path.join(os.path.dirname(os.path.realpath(__file__)), "pre_loaded_data", filename), sep=" ").astype(float)
+    df = pd.read_csv(os.path.join(os.path.dirname(os.path.realpath(__file__)), "pre_loaded_data", filename), sep=" ")
 
     if X_y:
         return torch.from_numpy(df.values[:, :-1]).float(), torch.from_numpy(df.values[:, -1]).float()
